Fix mysql2 wrapping
