+++ library/package.json
@@ -97,7 +97,8 @@
     "mongodb-v5": "npm:mongodb@^5.0.0",
     "mongodb-v6": "npm:mongodb@^6.0.0",
     "mysql": "^2.18.1",
-    "mysql2": "^3.10.0",
+    "mysql2-v3.10": "npm:mysql2@3.10",
+    "mysql2-v3.12": "npm:mysql2@3.12",
     "needle": "^3.3.1",
     "node-fetch": "^2",
     "percentile": "^1.6.0",
