+++ library/sinks/MySQL2.v3.10.test.ts
@@ -1,0 +1,3 @@
+import { createMySQL2Tests } from "./MySQL2.tests";
+
+createMySQL2Tests("mysql2-v3.10");
