+++ library/tsconfig.build.json
@@ -5,5 +5,5 @@
     "noEmit": false
   },
   "include": ["**/*.ts"],
-  "exclude": ["**/*.test.ts"]
+  "exclude": ["**/*.test.ts", "**/*.tests.ts"]
 }
